% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_surv_tree.R
\name{sim_surv_xgb_tree}
\alias{sim_surv_xgb_tree}
\title{Simple survival tree xgb}
\usage{
sim_surv_xgb_tree(model, x_data, y_data, top_n = NULL)
}
\arguments{
\item{model}{xgboot model object}

\item{x_data}{X data set}

\item{y_data}{Y data set}
}
\description{
This function allows you to find risk level automatically with lgbcv object.
}
\examples{
sim_surv_xgb_tree(model,x_data,y_data)
}
