% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xsurv.R
\name{Xsurv}
\alias{Xsurv}
\title{This function can fit the survival data with different model including xgboost,lightgbm,random forests, and gbm.}
\usage{
Xsurv(
  datax,
  datay,
  top_n = NULL,
  option = c("defaut", "xgb", "lgb", "gbm", "rf"),
  method = c("defaut", "pl", "C"),
  nfolds = 5,
  nround = NULL,
  lambda = NULL,
  alpha = NULL,
  eta = NULL,
  early_stopping_rounds = NULL,
  gtree = NULL,
  ncores = NULL,
  gfrac = NULL,
  gsh = NULL,
  rfnsp = NULL
)
}
\arguments{
\item{datax}{X data set}

\item{datay}{Y data set including time and event status}

\item{top_n}{number of top features for survival tree fitting(avaliable for xgb and lgb)}

\item{option}{model fitting option,defaut is xgb}

\item{method}{fitting metohd,defaut is 'pl' means using loss function:coxph likelihood}

\item{nfolds}{number of folds for crossvalidation}

\item{lambda}{l1 penalty parameter}

\item{alpha}{l2 penalty parameter}

\item{eta}{learning rate}

\item{early_stopping_rounds}{force a stopping round}

\item{gtree}{number of trees using in gbm model}

\item{ncores}{number of cores using in gbm}

\item{gfrac}{bag fraction in gbm with defaut 0.5}

\item{gsh}{shrinkage paramter in gbm with defaut 0.001}

\item{rfnsp}{number of splits in random forests}

\item{number}{of rounds}
}
\description{
This function can fit the survival data with different model including xgboost,lightgbm,random forests, and gbm.
}
\examples{
sim_surv_xgb_tree(model,x_data,y_data)
}
