% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsurv_predict.R
\name{xsurv_predict}
\alias{xsurv_predict}
\title{This function allows you to transform survival prediction from risk score to survival time.}
\usage{
xsurv_predict(model, x_train, y_train, x_test, nc = 25, nq = 0.9)
}
\arguments{
\item{model}{xgboost or lightgbm model object}

\item{x_train}{X train data set}

\item{y_train}{Y trian data set}

\item{x_test}{X test data set}

\item{nc}{number of cutting points defaut is 25, and 20-30 is recommended}

\item{nq}{quantile of train data time in each cut group for prediction time,defaut is and 0.9.0.8-0.95 is recommended.}
}
\description{
This function allows you to transform survival prediction from risk score to survival time.
}
\examples{
xsurv_predict(model,x_train,y_train,x_test)
}
